Extract peices of logic into seperate classes
